pyproject.toml
[project]
name = "nicknamer"
version = "2.0.0"
description = "Bot to facilitate nickname information management and viewing"
authors = [
    {name = "Andrew Cook", email = "acmagee@gmail.com"},
    {name = "Dušan Vučković", email = "dusanvucvr@gmail.com"}
]
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "discord.py (>=2.1,<3.0)",
    "PyYAML (>=6.0,<7.0)",
    "unalix (>=0.9,<1.0)",
    "urlextract (>=1.8,<2.0)",
    "dacite (>=1.9,<2.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
dpytest = "^0.7.0"
